refactor(stream): Generalize raw's name
