+++ crates/anstream/src/stream.rs
